Start to add database to store webhook deliveries
- Make `Database` generic
- Move config migrations into a new `ConfigDatabase`
- Start to add `WebhookDatabase`

Since GitHub Apps can now lookup webhook deliveries and request
redelivery via the api, we should start storing successfully
processed webhooks guids.

We can then use this to decide how far back to look in the webhook
delivery history to look for failures.
