+++ octobot/tests/pr_merge_test.rs
@@ -8,7 +8,7 @@ use tempdir::TempDir;
 use git_helper::temp_git::TempGit;
 use mocks::mock_github::MockGithub;
 use octobot_lib::config::Config;
-use octobot_lib::db::Database;
+use octobot_lib::config_db::ConfigDatabase;
 use octobot_lib::github;
 use octobot_lib::repos;
 use octobot_ops::pr_merge;
@@ -28,7 +28,7 @@ struct PRMergeTest {
 fn new_test() -> (PRMergeTest, TempDir) {
     let temp_dir = TempDir::new("repos.rs").unwrap();
     let db_file = temp_dir.path().join("db.sqlite3");
-    let db = Database::new(&db_file.to_string_lossy()).expect("create temp database");
+    let db = ConfigDatabase::new(&db_file.to_string_lossy()).expect("create temp database");
 
     let config = Arc::new(Config::new(db));
     config
