+++ ops/src/lib.rs
@@ -12,4 +12,6 @@ pub mod pr_merge;
 pub mod repo_version;
 pub mod slack;
 pub mod util;
+pub mod webhook_db;
+mod webhook_db_migrations;
 pub mod worker;
